Feature: targeting box and dropdown menu on click
In order to improve code readability, several components were made.

The ImageContainer renders the image, targeting box and dropdown menu.
It defines some state to re-render the targeting box and dropdown menu
each time the user clicks the image in a different position. It uses a
function, called when clicked to do this. And it's passed to the
TargetingBox to avoid code repetition.

The TargetingBox displays the box around the place the user made a
click. It receives the function, as wrote previously and the state, both
defined on ImageContainer component. It uses the style property to
update the position and visibility properties for each click.

The CharactersMenu contains a function that calculates the horizontal
position where the menu should be placed for each click. It also defines
an array of character names to reduce code thanks to certain React's
rendering technique.

Each component uses CSS modules to prevent certain selectors affecting
the whole page. TargetingBox and CharactersMenu use CSS variables to
define the initial position and visibility of their elements and also
to update them in React.