Add: character_positions db
Created a db that stores entries with the character name and the x and y
ranges in number that represent the position in pixels in the imgage.
These entries will be used to validate if the user found a character
when he clicks the image and selects a character.

Also created a controller, model and view. Finally, placed a route as
placeholder.