Refactor: boxStyle state variable
This state variable had object keys that refered to coordinates. Now
they refer to the property where they're applied in CSS.

A change made in ImageContainer component is the removal of the offset
variable in "onclick" function and made it another state variable.
This change was made to pass it as prop to CharactersMenu component. The
TargetingBox component was changed accordingly to these changes.

In CharactersMenu component, it accepts two new props:
characterPositions and offset. The offset prop use was explained last
paragraph. For characterPositions, it will be used later to get the
coordinates of a selected character and validate the player's clicks.

Another addition to CharactersMenu is several functions that help with
the validation said previously. They help to avoid repeating certain
lines of code. And also to know the steps to achieve this future
feature.

Lastly, added some temporarty comments to keep some to-dos for next
features.