Feature: play the game
Now the game can be played. It gives feedback when the player founds
a character or makes and error. To do this, the backend and frontend
are now "connected"; also, there are many changes. Some are new
functions, others are modifications in certain configuration files.

For the backend, there is a new gem installed: "rack-cors". This gem is
being added to allow fetch requests from the frontend. Also there is
some configuration in a new file to make this happen.

There are a few changes in the character_positions Controller and routes
file to prepare the data that will be fetched.

Now, for the frontend there are more changes or additions. First, there
is a new file that contains a custom hook, used to fetch the data from
the backend and be used in this project. It's contained in a file to
make other files more readable.

Another change is that App.jsx now has some conditional logic to render
an error or loading message. This is done thanks to the custom hook
described earlier, and also lets the user know if something happened
with the fetched data.

For the ImageContainer component there are other changes. There are new
state variables. These variables help render markers, the error message
or delete both the TargetingBox and CharactersMenu components. Some of
these variables are passed to CharactersMenu component to update its
values.

The CharactersMenu contains many new functions. These are used for
different reasons. Some update the state variables described in the
previous paragraph. Another ones determine the new values for these
state variables. One is to make this operations when the user selects
a character from the menu list.

Another change made to CharactersMenu is changed the props declaration.
Now the component mentions "props" instead of the actual props used
because it receives many props and makes the declaration shorter.

Other change in CharactersMenu is the creation of an array called
hiddenCharacters. This array is used to dynamically render the remaining
characters in the game and lets the player know how many are left to
find. A minor change is that some comments were removed.

Lastly, there are new lines in the CSS file of ImageContainer that adds
style to both markers and error message. And in the CSS file of
CharactersMenu was moved a background-color property from the ul to
the pop-up class to fix a minor error.