Feature: time user's round and display high scores
One of the most important parts to make these features possible is the
addition of Action Cable on Ruby on Rails. With that, this application
can use websocket connections, used to send data between both back end
end front end.

Also, to keep data separate in back end, there is now a new table,
called Scores, to store the best scores.

In the front end, there are new components, all of them are used to
prevent getting the App component larger, each one doing its part and
make the code easy to read.

For the new Modal component, it's done so it can be reused in the future
if needed.

Due to new components, the App component now has new state and context
variables, new methods and useEffects used to make the new features
possible.

Lastly, there are some changes in ImageContainer component. There is a
new state variable and useEffect to prevent a bug where the send method
from the websocket clients gets called many times. The other is a new
method attached to an Onload property in the div that contains the
image. This addition is to start timing properly when the user's device
loads the image.