Refactor: logic for rendering components
With new elements and styles added, the logic to render components was
changed. Before, it was using clientX / clientY properties from the
MouseEvent to calculate the TargetingBox and CharactersMenu positions.
Now it uses pageX / PageY properties from the same event.

It was also changed the entries of CharacterPositions to make the
coordinates based on the image coordinates. This was done because before
there were custom coordinates based on the page position.

Another change is the way it calculates the click properties. This is
also changed to match the change in the CharacterPositions entries. It
now uses the clientX / clientY properties from the MouseEvent and
subtracts the x / y properties from the DOMRect returned by
getBoundingClientRect method.

Other addition is the validation of coordinates before rendering the
TargetingBox and CharacterPositions components. This is done to prevent
a bug where the user can click near one of the image edges and then
click part of the TargetingBox outside the image to move it outside of
said image.

Lastly, minor changes are modifications on README.md and a new span
added to Stopwatch component.