Fix: markers wrongly placed when resizing window
To fix this bug are needed two changes: the first one is the div that
contains the image. this div was changed its css position value from
static to relative. Now the markers are positioned "relative" to this
container, instead of the window. The other change is the calculations
made to get the coordinates for the Targeting Box component. Previously,
the formula was (page coordinate - offset).
The new formula is: (page coordinate - image rect coordinate - window
scroll - offset). This ensures that the Targeting box, Characters Menu
and error renders correctly.

Another changes are the addition of a footer section to the app. Also
the stopwatch and high scores components have new fonts. Speaking of
high scores component, this one now renders between the before mentioned
stopwatch and the image container components. Lastly, there are minor
changes to README.MD.